#VJOCqmGY: Paws button style separation and themeability added
